Project and dataset sample
Project created to reproduce atlas problem, when, sometimes, duplicates first page

Select three lines in canvas, open "A3_ATLAS (selection) [with basemap]" composer, generate PDF. Most of the times, when Atlas PDF is created, the first page is duplcated

The problem has not been reproduced with "A3_ATLAS (selection)"
The problem has not been reproduced with "A3_ATLAS (selection) [with basemap]" when exported to image. Once the atlas has been exprted to images, when exporting to PDF, the problem does not appear

Aditional problem: Layer "atlas_fail", style "Atlas".

Opacity of the symbol is controled by selection. If line is selected, line is opaque (should be shown). Lines not selected are transparent (opacity = 0).
Line symbol opacity is correct, but start and point, generated by geometry generator, are not.